Get current time with timezone awareness

Reports the current timestamp in UTC and optionally in a local or specified timezone. Uses a detection cascade to find your local timezone when the system reports UTC.

timenow(tz = NULL, quiet = FALSE)

Arguments

tz

Optional timezone specification. Can be:

NULL (default): show UTC and detected local timezone

A valid IANA timezone name: "Australia/Hobart"

A fuzzy string: "Perth Australia", "new york", "tokyo"

quiet

Suppress the local timezone detection message

Value

A named list of class "timenow" with UTC and local times (invisibly

Details

Timezone detection cascade

When your system timezone is UTC (common on servers), timenow uses this cascade to find your actual local timezone:

getOption("timenow.tz") - set in .Rprofile

Sys.getenv("R_TIMENOW_TZ") - set in .Renviron

/etc/timezone file (Debian/Ubuntu)

timedatectl output (systemd)

Sys.timezone() - the R default

Setting your timezone

In ~/.Renviron:

R_TIMENOW_TZ=Australia/Hobart

Or in ~/.Rprofile:

options(timenow.tz = "Australia/Hobart")

Examples

timenow()
#> 2026-02-03 11:09:13 (UTC)
#> 2026-02-03 11:09:13 (Etc/UTC)
#> same as UTC
timenow("Perth")
#> 2026-02-03 11:09:13 (UTC)
#> 2026-02-03 19:09:13 (Australia/Perth)
#> +8h from UTC
timenow("US Eastern")
#> 2026-02-03 11:09:13 (UTC)
#> 2026-02-03 06:09:13 (US/Eastern)
#> -5h from UTC